Refactor: Modularize codebase for improved organization
This commit completes the initial refactoring phase aimed at improving the structure and organization of the 2.5D Lands game.

Key changes:
- Decomposed the monolithic `main.cpp` into several focused modules:
    - `GameConstants.h`: Centralized all global game constants.
    - `DataTypes.h`: Consolidated all primary data structures (Tile, Player, Camera, BiomeProperties, TerrainData) and enums (BiomeType).
    - `Noise.h` / `Noise.cpp`: Encapsulated Perlin noise generation logic and associated helper functions.
    - `World.h` / `World.cpp`: Grouped world generation functions, including terrain height calculation, biome determination, and the main `GenerateWorld` function.
    - `Renderer.h` / `Renderer.cpp`: Separated rendering-specific tasks such as `RenderTile`, `RenderPlayer`, and the `WorldToScreen` coordinate transformation.
    - `Player.h` / `Player.cpp`: Isolated player-related logic, specifically the `HandlePlayerMovement` function.
    - `Camera.cpp`: Provided the implementation for the `Camera::update` method (its declaration remains in `DataTypes.h`).
- Updated `CMakeLists.txt` to include all new source files (`.cpp`) in the compilation process.
- Successfully compiled the entire project after each refactoring step to ensure integrity.
- Installed necessary SDL2 development libraries in the build environment to enable successful compilation.

The `main.cpp` file is now significantly streamlined, primarily responsible for:
- SDL initialization and window/renderer management.
- The main game loop.
- Event handling.

This modular structure provides a more organized and maintainable codebase, paving the way for the implementation of new features as outlined in the project plan.
